Making the permissions controller use a behavior so we can extend the behavior in other gems
